Safely interpose fcntl() on architectures with bounded variadic arguments
`fcntl()` abuses C variadic arguments to implement optional parameters.
We want to forward this argument to the real `fcntl()` call, but doing
so is non-portable. On most architectures, we can just read an
`intptr_t`/`void*` and this will give us a (semi-)valid result even if
no argument/an int was passed instead since `va_arg()` will generally
read the next argument register or arbitrary data from the stack.

On CHERI-enabled architectures, variadic arguments are tightly
bounded, which means that reading a `void*` if an `int` was passed will
result in a runtime trap. This commit adds a switch ensures that we read
the correct number of bytes from the variadic arguments.
This ugly workaround would not be needed if a vfcntl() call taking
a va_list existed, or if fcntl just took an intptr_t argument.

This reduces the number of test failures on CheriBSD Morello from 210
to 8.